# services/mapi-xdg/docs/indexer/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dogecoin - Blockchain Indexer API",
    "description": "Dogecoin indexer endpoints: addresses, dunes, inscriptions and DRC-20. The same binary also proxies a set of node RPC endpoints (blocks, mempool, chain info) documented in the repository README.",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.txt"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local instance"
    }
  ],
  "paths": {
...
    "/addresses/{address}/transfer_inscriptions": {
      "get": {
        "tags": [
          "Addresses"
        ],
        "summary": "DRC20 Transfer Inscriptions by Address",
        "description": "List of all unused transfer inscriptions which reside at the specified address or script pubkey.",
        "operationId": "transfer_inscriptions_by_address",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "Dogecoin address or hex encoded script pubkey",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "DNG3G7pKc1DgciyWs36GFUnVyvJieDhKdb"
          },
          {
            "name": "ticker",
            "in": "query",
            "description": "DRC20 ticker string",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "example": "TUAH"
          },
          {
            "name": "count",
            "in": "query",
            "description": "The max number of results per page",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "type": "integer",
                  "default": 100,
                  "minimum": 0
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "The order in which the results are sorted",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "type": "string",
                  "default": "asc",
                  "enum": [
                    "asc",
                    "desc"
                  ]
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Pagination cursor string, use the cursor included in a page of results to fetch the next page",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Requested data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTransferInscriptionByAddress"
                },
                "example": {
                  "data": [
                    {
                      "inscription_id": "f02da3d6bebab13d5d604be1ed73d9a9c677dadf6ca71bc5fff7d99cdead11b0i0",
                      "sat_amount": 5678,
                      "ticker": "TUAH",
                      "token_amount": "1234",
                      "utxo_block_height": 843010,
                      "utxo_confirmations": 23700,
                      "utxo_sat_offset": 0,
                      "utxo_txid": "e2283e7c915ef074806136e0002cbc69f5fdd2e9f70f14b0eab48cdcbe867cc1",
                      "utxo_vout": 0
                    }
                  ],
                  "last_updated": {
                    "block_hash": "1c3e7cd9e46bd6d0adb1ae0b52ec1a1ddfaa0cb61a41a1b1c26f0d1b0f42a06c",
                    "block_height": 5146710
                  },
                  "next_cursor": "BFRVQUhgfQot2JciKRPVj8lXsEKVJhF6CmHJZGQv6TsHfzKMzsEAAAAA"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameters"
          },
          "404": {
            "description": "Requested entity not found on-chain"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LastUpdated": {
        "type": "object",
        "required": [
          "block_hash",
          "block_height"
        ],
        "properties": {
          "block_hash": {
            "type": "string"
          },
          "block_height": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "PaginatedTransferInscriptionByAddress": {
        "type": "object",
        "required": [
          "data",
          "last_updated"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransferInscriptionByAddress"
            }
          },
          "last_updated": {
            "$ref": "#/components/schemas/LastUpdated"
          },
          "next_cursor": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "TransferInscriptionByAddress": {
        "type": "object",
        "required": [
          "ticker",
          "inscription_id",
          "token_amount",
          "sat_amount",
          "utxo_txid",
          "utxo_vout",
          "utxo_sat_offset",
          "utxo_block_height",
          "utxo_confirmations"
        ],
        "properties": {
          "inscription_id": {
            "type": "string",
            "description": "String representation of an inscription ID, whose first coordinate is the reveal\ntransaction hash, and the second coordinate is the index of the new inscription."
          },
          "sat_amount": {
            "type": "integer",
            "format": "int64",
            "description": "Number of sats locked in the UTxO.",
            "minimum": 0
          },
          "ticker": {
            "type": "string"
          },
          "token_amount": {
            "type": "string",
            "description": "Number of tokens locked in the UTxO."
          },
          "utxo_block_height": {
            "type": "integer",
            "format": "int64",
            "description": "Block height of the UTxO containing the inscription.",
            "minimum": 0
          },
          "utxo_confirmations": {
            "type": "integer",
            "format": "int64",
            "description": "Number of confirmations of the block where the UTxO containing the inscription was created.",
            "minimum": 0
          },
          "utxo_sat_offset": {
            "type": "integer",
            "format": "int32",
            "description": "Offset of inscribed sat in the UTxO containing it.",
            "minimum": 0
          },
          "utxo_txid": {
            "type": "string",
            "description": "Tx ID of the UTxO containing the inscription."
          },
          "utxo_vout": {
            "type": "integer",
            "format": "int32",
            "description": "Output index of the UTxO containing the inscription.",
            "minimum": 0
          }
        }
      },
...
    }
  }
}